pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "larksh"
version = "0.1.0"
description = "通过飞书机器人控制内网服务器 Shell 的工具"
requires-python = ">=3.10"
dependencies = [
    "lark-oapi>=1.3.0",
    "ptyprocess>=0.7.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.0.0",
]

[project.scripts]
larksh = "main:main"

[tool.setuptools.packages.find]
exclude = ["tools*", "docs*", "dist*"]
